backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart
